Implements two versions of a calculator: Infix & Postfix Notation.
Implements two versions of a calculator: Infix & Postfix Notation in the Evaluator.java file. Also, for this assignment, the input is assumed to be correct, and we did not have to test for 'garbage' input. The input will also have a space between every entry.
The PDF document demonstrates test cases for the program.